+++ automad/src/server/App.php
@@ -56,7 +56,7 @@ defined('AUTOMAD') or die('Direct access not permitted!');
  * @license See LICENSE.md for license information
  */
 class App {
-	const VERSION = '2.0.0-beta.39';
+	const VERSION = '2.0.0-beta.40';
 
 	/**
 	 * Required PHP version.